requirements.txt
colorama==0.4.6
DAWG-Python==0.7.2
docx2pdf==0.1.8
et_xmlfile==2.0.0
lxml==5.3.0
num-to-rus==1.1.0
openpyxl==3.1.4
pymorphy3==2.0.1
pymorphy3-dicts-ru==2.4.417150.4580142
python-docx==1.1.2
pytrovich==0.0.4
pywin32==308
tqdm==4.67.0
typing_extensions==4.12.2
